requirements.txt
asgiref==3.11.0
Django==5.2.8
django-cors-headers==4.9.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.15.0
